add new hit sound for asteroids
-and temporarily disable terrible f3 sound until I can replace with something less silly. it doesn't sound good. (why did I leave it so long? because I had to fix engine things before making more assets. also I am not great at sound design)
-make shooting a little quieter
-rename to cannon because it's not the laser
-couple notes

+++ core/src/com/spaceproject/components/StatsComponent.java
@@ -5,6 +5,7 @@ import com.badlogic.ashley.core.Component;
 public class StatsComponent implements Component {
 
     //stats
+    //public long timeAlive;
     //todo: kills, deaths, resourcesCollected, creditsEarned, creditsSpent, etc...
     // should it collect per life and total on death?
     public long distanceTraveled;
@@ -17,11 +18,13 @@ public class StatsComponent implements Component {
 
     public long shotsHit;
 
-    public long damageDealt;
+    public float damageDealt;
 
-    public long damageTaken;
+    public float damageTaken;
 
     public long resourcesCollected;
+
+    //public long resourcesSpent;
 
     public long resourcesLost;
